requirements_build.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
itsdangerous==2.1.2
requests==2.31.0
gunicorn==21.2.0
pyinstaller==6.3.0
